Fixed milestone release auto title and consolidate workflows
Workflows that are triggered by the same event (PR labeler, PR to
Milestone and PR title) are now in the same reusable workflow. This
makes the sharing of infos between jobs easier and also the maintenance
of the jobs.

This also fixed an issue with  the milestone PR auto title. Because the
PR was attached to the milestone during the same github event as the PR
auto title, the `pull_request.milestone` was null in the event for the
PR auto title. Therefore this workflow could not find out the correct
next version. To solve this we need to have the PR to Milestone Workflow
together with the PR auto title and pass the milestone value to the Bump
Milestone action (requier v1.2.0).

Added also a name to each job. This name is then displayed in github
when the job is runned inside a PR.