Changed hotfix release notes
The hotfix PR title where prefixed with the version. Because the hotfix
PR title are then added to the Release Notes, this makes the version in
RN redundant and not very nice. Therefore now we don't prefix anymore
the hotfix PR with the version and we also labeled them as hotfix to
specify it in RN.