Fixed bump type when taken from git commit message
When a git commit message contained a bump type (e.g. #major, #minor or #patch)
the geoadmin/github-tag-action@master used it to bump the tag, but the title
was set with the bump type taken from the `bump_type` action script which only
looked at the PR title, body and branch name. This action is actually only
used to defined the default bump type and the github-tag-action give the
actual bump type.