Fixed a bug in the `Add PR to Milestone` when no milestone exist
In the `Add PR to Milestone` action, when the milestone did not exists (e.g.
has been already closed or deleted) the action failed silently. In this
case the version in the PR title might have been wrongly set. This happened
at least once.

Now this action is failing in case the milestone doesn't exists yet and
other actions are not performed.