Auto milestone branch protection and deletion
Now when a milestone branch is created, a branch protection for this one is
automatically created. The protection requires at least one approval and
prevent merging directly to the branch. With the workflow input ci_status_check_name
we can configure a CI status check that must path in order to merge. It also
set the default branch to the created milestone branch.

When a milestone branch is merged and/or closed, the branch protection is
automatically removed and the branch is deleted. The default branch is set
back to master.