Milestone workflow bug - update milestone branch with hotfix
If the milestone branch is protected in order to only allow merge of branch
that are up to date (aka "Require branches to be up to date before merging."),
then we cannot merge `master` into `develop-YYYY-MM-DD`.

The least worse solution for this problem is to allow merging non up to date
branch into the milestone branch. We can leverage the problem by setting the
general setting of the repo with "Always suggest updating pull request branches "
so when a regular feature branch PR to milestone is created the user will be
notify if its branch is not up to date, although he could still merge his PR
even his branch is not up to date.