Fixed the PR auto labeller bug
If a PR had a head branch with `data-` or `data_` in its name (regardless of its
position in the name, the `data-integration` label was added.

Now the regex are correct and only check for the start of the string. With
the exception for the `skip-release-note` label that is added even if the
`norn`, `skip-rn`, `skiprn` text is found in the head branch.