Fix issue with bump type in title being overwritten
The BitJonny/gh-get-current-pr@3.0.0 action had an issue with pull_request events,
somehow it returned the last closed PR on the head branch instead of the curren t
PR on the head.

Because the PR infos required are anyway part of the github pull_request event
payload we take it from it now.