Allow to overwrite previous bump_type
If any commit in a PR contains a bump type it was taken instead of the PR
bump type set in title or body. This was due to the fact that geoadmin/github-tag-action
takes bump type from commit message and we were setting the bump type from the
title as default.

So now the PR bump type is set as commit message to the local merge commit that
is used by geoadmin/github-tag-action to compute the future git tag and bump type.