PB-117: Allow to skip a release (RN and tagging) based on commit message
Now if the last commit contains #skip-tagging-and-release then no release will be created.

This is needed in order to be able to modify the release.yml configuration file
that is managed by terraform and directly push on master branch. Without this
the modification of this configuration file for generating release note would
trigger a new empty release.

NOTE: this is only needed for semver because the milestone release are triggered
on PR closed event and not push even like the semver.